Moved TestClassloader into its own java file. Was causing test plugin to die.
When this was fixed, over 30 errors and 20 or so failures in unit tests
that had not been being run showed up.  Ant apparently stopped running
the tests, but still showed success.  For instance, the MultiLoaderTestCase
should not pass as written.  When using ant to build the original project,
the MultiLoaderTestCase surefire results do not exist in the test-reports
directory.