IncludeRelativePath was not called. Compare string not loaded. Helper func.
Turns out there was already a handler for relative paths, but it was not
being injected in all the test cases. Scanner was finding an end of file
token in one of the comparison test files, so made it a loop.  Also added
a helper function, calcPathToTestDirectory(), to determine the directory for the
FILE_RESOURCE_LOADER_PATH.