Fixed path issues and removed extra eol that was being eaten by Scanner.
Initially was using \\Z as an end token, but it was eating the last eol at
the end of the test comparison file.  Switched to \\z, which since it was
not eating it, now reverted changes to test files.  It looked like they
matched, but it took playing with several tests before I realized the test
files really did not match.