Something broke this test, but the code doesn't look like it should work.
Strangeness is that it works in the 1.7 source download with limited
changes.  I can't see where I changed something that would have caused the
instance being cast to have been replaced with an instance that implemented
the SecureIntrospectorControl interface.  So I am moving this to the side
for the moment.